Refactored the network state adapter. (#254)
