+++ app/src/main/java/com/guillermonegrete/gallery/folders/FolderAdapter.kt
@@ -6,9 +6,6 @@ import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.content.ContextCompat
-import androidx.core.view.isVisible
-import androidx.paging.LoadState
-import androidx.paging.LoadStateAdapter
 import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
@@ -17,7 +14,6 @@ import com.google.android.material.color.MaterialColors
 import com.guillermonegrete.gallery.R
 import com.guillermonegrete.gallery.databinding.FolderItemBinding
 import com.guillermonegrete.gallery.databinding.FolderNameItemBinding
-import com.guillermonegrete.gallery.databinding.ItemNetworkStateBinding
 import com.guillermonegrete.gallery.folders.models.FolderUI
 import io.reactivex.rxjava3.subjects.PublishSubject
 import kotlin.math.abs
@@ -172,33 +168,5 @@ object FolderDiffCallback : DiffUtil.ItemCallback<FolderUI>() {
 
     override fun areContentsTheSame(oldItem: FolderUI, newItem: FolderUI): Boolean {
         return oldItem == newItem
-    }
-}
-
-class NetworkStateAdapter(private val retry: () -> Unit) : LoadStateAdapter<NetworkStateAdapter.ViewHolder>() {
-
-    override fun onBindViewHolder(holder: ViewHolder, loadState: LoadState) {
-        holder.bind(loadState)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, loadState: LoadState): ViewHolder {
-        val binding = ItemNetworkStateBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return ViewHolder(binding, retry)
-    }
-
-    class ViewHolder(val binding: ItemNetworkStateBinding, retry: () -> Unit) : RecyclerView.ViewHolder(binding.root) {
-
-        init {
-            binding.retryButton.setOnClickListener { retry() }
-        }
-
-        fun bind(loadState: LoadState) = with(binding) {
-            progressBar.isVisible = loadState is LoadState.Loading
-
-            // Error views
-            retryButton.isVisible = loadState is LoadState.Error
-            errorMsg.isVisible = loadState is LoadState.Error
-            if (loadState is LoadState.Error) errorMsg.text = loadState.error.localizedMessage
-        }
     }
 }
